Experiment with REST-based /api
I was looking at tying the tables work into linking to each
individual resource when you painted the table. However, we don't
currently have an API which easily supports that, and we do have
'action' endpoints, which is appropriate for a frontend but not
necessarily a backend, API-based service.

So, what to do?

One thing I was thinking I could try is have the /api prefix serve
as the REST endpoints, and the /ui prefix or something serve as
a more traditional, form-based flow.